Fix transformation of HTML "for" attribute
The React docs say that not only the HTML "class" attribute becomes
"className" in JSX, but also "for" becomes "htmlFor" [1]. The first one
was already implemented in Kioo, the second not. Implement it.

Note that this is not tested, yet, since I couldn't get the build to run
on my system.

[1] https://facebook.github.io/react/docs/tags-and-attributes.html