Fix more HTML attributes
Apparently it's not only "class" and "for" that can't be converted
between JSX and HTML automatically, but other attributes as well.
Add accept-charset/acceptCharset to the manual conversions. Add
formEncType to the list of auto-convertible attributes. There are
probably some more things missing, but I don't have the nerve to go
through the HTML and JSX spec right now.